add create dictionary api integration test with unexpected errors occured

+++ DictionaryAPI/DictionaryAPI.Tests/Integration/DictionaryDbTestHelper.cs
@@ -6,6 +6,8 @@ public static class DictionaryDbTestHelper
 {
     public static void DeleteDb(Guid dbId)
     {
+        if (dbId == Guid.Empty) return;
+        
         var path = DictionaryDbPathProvider.GetPath(dbId);
         
         File.Delete(path);
